change some data type and field name, fix conflict

+++ src/main/java/org/petspa/petcaresystem/authenuser/service/AuthenUserService.java
@@ -7,7 +7,6 @@ import org.springframework.stereotype.Service;
 
 @Service
 public interface AuthenUserService {
-
-    ResponseEntity<ResponseObj> UpdateProflie(int id, UpdateProfileRequest profileRequest);
+    ResponseEntity<ResponseObj> UpdateProflie(Long id, UpdateProfileRequest profileRequest);
 
 }
